requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.23.0
python-dotenv==1.0.0
google-genai==0.3.0
